--- src/pages/test/style.css
@@ -1,21 +1,0 @@
-section {
-    flex: 1 0 auto;
-    position: relative;
-    width: 100%;
-    -webkit-transition: all .5s ease-in;
-    -o-transition: all .5s ease-in;
-    transition: all .5s ease-in;
-}
-
-.homeImage {
-    max-width: 100%;
-    width:max-content;
-    height: auto;
-    padding: 0;
-    margin: 0;
-    padding-bottom: 10px;
-}
-
-.ac_btn {
-    margin-bottom: 10px;
-}
